feat: target a minimum VMAF with an interpolated CRF search
Target quality now treats the configured VMAF as a hard minimum and keeps the highest CRF (lowest bitrate) that still holds it, instead of accepting any score within a tolerance band.

- search the 0.25 CRF grid with an interpolated binary search; stop on an early in-tolerance hit, an adjacent-step bracket, or the probe budget
- replace the fixed probe count and two-sided tolerance with min_probes/max_probes and a single above-floor tolerance
- add max_encoded_percent: cap a chunk against its actual source bytes, raising the CRF (and dropping below the floor with a warning) when holding quality would blow up the bitrate
- raise the CRF ceiling to 70 and pick fractional CRFs
- log every probe with its CRF, VMAF, and size, plus the final pick per chunk
- discard cached chunk CRFs when target_quality settings change, not just encoder settings

+++ src/encode.rs
@@ -10,9 +10,10 @@ use crate::ext::external_bin;
 use crate::resume::SceneEntry;
 
 /// Per-call CRF/preset overrides (target quality probes and final encode).
+/// CRF is fractional: the SVT-AV1 encoders accept 0.25 steps.
 #[derive(Default, Clone, Copy)]
 pub struct EncodeOverrides {
-    pub crf: Option<u32>,
+    pub crf: Option<f64>,
     pub preset: Option<u32>,
 }
